+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
  responsibility.
  * Consumer is now responsible for creating Invoker object.
  * Spork return a Proxy object that allows calling the right method on service object.
+ * New schema for `DoctrineDriver`. Queues are now kept in a seperate table.
 
 0.7.1 / 2013-07-12
 ------------------